# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alpha Cyber OSINT Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/search:
    post:
      operationId: startSearch
      tags: [tools]
      summary: Start an OSINT search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchInput"
      responses:
        "201":
          description: Search started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchJob"
  /searches:
    get:
      operationId: listSearches
      tags: [searches]
      summary: List search history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Search history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchPage"
  /searches/{searchId}:
    get:
      operationId: getSearch
      tags: [searches]
      summary: Get search details and results
      parameters:
        - name: searchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchDetails"
    delete:
      operationId: deleteSearch
      tags: [searches]
      summary: Delete a search record
      parameters:
        - name: searchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /searches/{searchId}/export:
    post:
      operationId: exportSearch
      tags: [searches]
      summary: Export search results
      parameters:
        - name: searchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Export URL or data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    SearchInput:
      type: object
      properties:
        target:
          type: string
          description: "Username, email, phone number, or full name"
        targetType:
          type: string
          enum: [username, email, phone, name, auto]
        tools:
          type: array
          items:
            type: string
          description: "List of tools to run, or empty for auto-select"
      required:
        - target
        - targetType
    SearchJob:
      type: object
      properties:
        searchId:
          type: string
        target:
          type: string
        targetType:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        tools:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - searchId
        - target
        - targetType
        - status
        - tools
        - createdAt
    ToolResult:
      type: object
      properties:
        tool:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        results:
          type: array
          items:
            $ref: "#/components/schemas/ResultItem"
        duration:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
      required:
        - tool
        - status
        - results
    ResultItem:
      type: object
      properties:
        platform:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        found:
          type: boolean
        metadata:
          type: ["string", "null"]
      required:
        - found
    SearchDetails:
      type: object
      properties:
        searchId:
          type: string
        target:
          type: string
        targetType:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        tools:
          type: array
          items:
            type: string
        toolResults:
          type: array
          items:
            $ref: "#/components/schemas/ToolResult"
        totalFound:
          type: integer
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - searchId
        - target
        - targetType
        - status
        - tools
        - toolResults
        - totalFound
        - createdAt
    SearchSummary:
      type: object
      properties:
        searchId:
          type: string
        target:
          type: string
        targetType:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        totalFound:
          type: integer
        toolCount:
          type: integer
        createdAt:
          type: string
      required:
        - searchId
        - target
        - targetType
        - status
        - totalFound
        - toolCount
        - createdAt
    SearchPage:
      type: object
      properties:
        searches:
          type: array
          items:
            $ref: "#/components/schemas/SearchSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - searches
        - total
        - page
        - limit
...
    ExportInput:
      type: object
      properties:
        format:
          type: string
          enum: [pdf, csv, json]
      required:
        - format
    ExportResult:
      type: object
      properties:
        format:
          type: string
        data:
          type: string
          description: "Base64 encoded data or download URL"
        filename:
          type: string
      required:
        - format
        - data
        - filename