# lib/api-zod/src/generated/types/searchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alpha Cyber OSINT Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchInputTargetType } from './searchInputTargetType';

export interface SearchInput {
  /** Username, email, phone number, or full name */
  target: string;
  targetType: SearchInputTargetType;
  /** List of tools to run, or empty for auto-select */
  tools?: string[];
}